pyproject.toml
[project]
name = "hierarchy"
version = "3.0.19"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.17.0",
    "asyncpg==0.30.0",
    "confluent-kafka[protobuf,schemaregistry]==2.12.0",
    "elasticsearch[async]==8.19.1",
    "fastapi==0.119",
    "grpcio==1.75.1",
    "protobuf==6.33.0",
    "pydantic==2.12.2",
    "pyjwt[crypto]==2.10.1",
    "python-keycloak==5.8.1",
    "sqlalchemy[asyncio]==2.0.44",
    "sqlalchemy-utils==0.42.0",
    "sqlmodel==0.0.27",
    "uvicorn[standard]==0.37.0",
    "pydantic-settings==2.11.0",
]

[dependency-groups]
dev = [
]
lint = [
    "ruff==0.14.0",
]
tests = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-mock==3.15.1",
    "httpx==0.28.1",
    "testcontainers==4.13.2",
    "pytest-cov==7.0.0",
]
security = [
    "pip-audit==2.9.0",
]
migrations = [
    "alembic>=1.17.0",
    "asyncpg==0.30.0",
    "grpcio>=1.75.1",
    "protobuf>=6.33.0",
    "pydantic-settings>=2.11.0",
    "requests>=2.31.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "sqlalchemy-utils>=0.42.0",
    "sqlmodel>=0.0.27",
]
